libmetacity: replace ints with doubles in few places
When we draw decorations we scale down client width and height. This
causes problem when client width or height is odd number - decorations
is not drawn in full size. Replace ints with doubles to make sure
that cairo scales decorations to correct size.

Improves HiDPI support.

+++ libmetacity/meta-gradient-spec.c
@@ -137,10 +137,10 @@ meta_gradient_spec_render (const MetaGradientSpec      *spec,
                            const MetaAlphaGradientSpec *alpha_spec,
                            cairo_t                     *cr,
                            GtkStyleContext             *context,
-                           gint                         x,
-                           gint                         y,
-                           gint                         width,
-                           gint                         height)
+                           gdouble                      x,
+                           gdouble                      y,
+                           gdouble                      width,
+                           gdouble                      height)
 {
   cairo_pattern_t *pattern;
 
@@ -224,10 +224,10 @@ void
 meta_alpha_gradient_spec_render (MetaAlphaGradientSpec *spec,
                                  GdkRGBA                color,
                                  cairo_t               *cr,
-                                 gint                   x,
-                                 gint                   y,
-                                 gint                   width,
-                                 gint                   height)
+                                 gdouble                x,
+                                 gdouble                y,
+                                 gdouble                width,
+                                 gdouble                height)
 {
   if (!spec || spec->n_alphas == 1)
     {
